added fix to trait: CuratorServiceIntegration which was creating an httpclient whose constructror has changed. The use of the old constructor caused a failur to initialize any object that extends the trait, causing a difficult to understand error.
 updated geojson library to rojoma-json-v3

fixed a few of the decode cases to show how the decoding error cases pass through.

used right projections with map to handle the right success and left failure cases from decoding in a more compact way.

updated geoJsonTest to rojoma-json-v3

fixed GeoJsonTest, next is to finish the JtsCodecsTest

last test for geojson changes to rojoma-json-v3, note tests were changed to account for the new Either, functionality that did not exist before.

updated the tests that I could, and have set it up for future tests to use arbitrary coordinates and other geometry objects

changed boundaries for abitrary coordinate creation for geojson tests  and cleaned up a bit

updated all the tests to use arbitrary, with chi's suggestion that we use gen.choose as opposed to arbitrary.double since it would fail to satisfy constraints and therefore fail too often.

reduced limits to avoid hitting them for testing, code coverage will not be impacted significantly.